// src/app/panel3/shared/interfaces/user.interface.ts
export interface UserInterface {
  idcompany: number;
  externalid: number;
  payeeid: number;
  title: string;
  surname: string;
  firstname: string;
  middlename: string;
  preferredname: string;
  gender: string;
  birthdate: Date;
  email: string;
  mobilephone: string;
  homephone: string;
  workphone: string;
  idaddress: number;
  idpostaladdress: number;
  emergencycont: string;
  emergencynumb: string;
  iduser: number;
  enable: string;
  id: number;
}